fix: update relative paths to hooks/api/utils/context for moved components
