+++ frontend/src/components/crawl/CrawlDialsPanel.jsx
@@ -15,7 +15,7 @@
 import React, { useState } from "react";
 import { Settings2, ChevronDown, ChevronUp } from "lucide-react";
 import TestDials from "../shared/TestDials.jsx";
-import { countActiveDials, loadSavedConfig } from "../utils/testDialsStorage.js";
+import { countActiveDials, loadSavedConfig } from "../../utils/testDialsStorage.js";
 
 export default function CrawlDialsPanel({ value, onChange }) {
   // Controlled mode: parent owns the config via value/onChange.
