+++ frontend/src/components/layout/ProtectedRoute.jsx
@@ -21,7 +21,7 @@
  */
 
 import { Navigate, useLocation } from "react-router-dom";
-import { useAuth } from "../context/AuthContext.jsx";
+import { useAuth } from "../../context/AuthContext.jsx";
 
 function AuthLoadingSkeleton() {
   return (
