+++ frontend/src/components/run/RunRegressionModal.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { Play, X, RefreshCw } from "lucide-react";
-import { api } from "../api.js";
+import { api } from "../../api.js";
 import ModalShell from "../shared/ModalShell.jsx";
 
 /**
